backend/requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python-headless
tensorflow
numpy
pillow
python-multipart
matplotlib